+++ src/expression.cpp
@@ -267,7 +267,7 @@ void Diagram::binarize(Expression& expr) const
     //recursive calls of transform2Expr inside depending on "relations" of residual tensor bt
     transform2Expr(expr,inters,order,bt);
   }
-  for ( TensorsSet::iterator it = expr._tensors.begin(); it != expr._tensors.end(); ++it) {
+  for ( TensorsList::iterator it = expr._tensors.begin(); it != expr._tensors.end(); ++it) {
     if ( it->equal(res) ) {
       const Tensor * pRes = &(*it);
       expr.addresidual(pRes);
@@ -395,7 +395,7 @@ const SlotType* Expression::add(const SlotType& slottype)
 
 const Tensor* Expression::add(const Tensor& tensor)
 {
-  for ( TensorsSet::iterator it = _tensors.begin(); it != _tensors.end(); ++it) {
+  for ( TensorsList::iterator it = _tensors.begin(); it != _tensors.end(); ++it) {
     if ( it->equal(tensor) ) {
       return &(*it);
     }
@@ -432,7 +432,7 @@ const Action* Expression::add(const Action* pAction)
 
 const Tensor* Expression::add2residual(const Tensor& res, const Action * pAct)
 {
-  for ( TensorsSet::iterator it = _tensors.begin(); it != _tensors.end(); ++it) {
+  for ( TensorsList::iterator it = _tensors.begin(); it != _tensors.end(); ++it) {
     if ( it->equal(res) ) {
       it->add(pAct);
       return &(*it);
@@ -444,7 +444,7 @@ const Tensor* Expression::add2residual(const Tensor& res, const Action * pAct)
 
 const Tensor* Expression::find(const Tensor& tensor, bool considerprops) const
 {
-  for ( TensorsSet::const_iterator it = _tensors.begin(); it != _tensors.end(); ++it)
+  for ( TensorsList::const_iterator it = _tensors.begin(); it != _tensors.end(); ++it)
     if ( it->equal(tensor,considerprops) ) return &(*it);
   return 0;
 }
@@ -756,7 +756,7 @@ std::ostream & operator << (std::ostream& o, const Expression& exp) {
     o << st << std::endl;
   }
   // tensors....
-  const TensorsSet& ts = exp.tensors();
+  const TensorsList& ts = exp.tensors();
   std::set<Tensor> uniquetensortypes;
   for (const auto& t: ts){
     if( (uniquetensortypes.insert(t)).second)